requirements.txt
asyncio
python-dotenv
gradio_client
